CansatEmulator/requirements.txt
pynput
pyserial